expect unpoisoned mutex
Realistically, the lock method only returns an error if another thread panicked while holding the mutex. This could only happen if te init function panicked, which would only happen if there was a bug in the windows-rs crate, in which case there are greater issues than this specific function panicking.